introduce long weak links
Quite disruptive, but it needed doing. Now need to think quite a lot about
overlap.

One thing we're not doing yet is to favor short links to long links when
possible.

The non-overlapping() routine is a bit greedy, and may end up with a local
optimum instead of the best possible set of links.

We're also not tackling the beast of several possible ways to form a group yet.